+++ consensus-proof/build.rs
@@ -1,6 +1,5 @@
 extern crate prost_build;
 
 fn main() {
-    println!("hello from build.rs");
-    prost_build::compile_protos(&["src/milestone.proto"], &["src/"]).unwrap();
+    prost_build::compile_protos(&["src/types.proto"], &["src/"]).unwrap();
 }
